pyproject.toml
[project]
name = "ds-resume-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-levenshtein>=0.27.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "thefuzz>=0.22.1",
    "tqdm>=4.67.2",
]
